// frontend/src/openapi/models/req-signup-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ReadersLounge API
 * ReadersLounge API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ReqSignupBody
 */
export interface ReqSignupBody {
    /**
     * 
     * @type {string}
     * @memberof ReqSignupBody
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof ReqSignupBody
     */
    'identifier'?: string;
    /**
     * 
     * @type {string}
     * @memberof ReqSignupBody
     */
    'credential'?: string;
}